gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx>=0.23.0,<0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
starlette==0.27.0
email-validator==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
itsdangerous==2.1.2
supabase==1.0.3  # Supabase Python client